+++ nrfutil/device/common.ts
@@ -232,7 +232,11 @@ export const getDeviceSandbox = async () => {
             }
         });
 
-        deviceSandbox.setLogLevel(getIsLoggingVerbose() ? 'trace' : 'error');
+        const fallbackLevel =
+            process.env.NODE_ENV === 'production' ? 'off' : 'error';
+        deviceSandbox.setLogLevel(
+            getIsLoggingVerbose() ? 'trace' : fallbackLevel
+        );
         // Only the first reset after selecting "reset with verbose logging" is relevant
         persistIsLoggingVerbose(false);
     }